+++ NumInWord/TensUnit.cs
@@ -2,7 +2,7 @@
 {
     internal class TensUnit : Wordable
     {
-        private int unitVal;
+        private long unitVal;
 
         public TensUnit() {
             priority = 2;
@@ -12,13 +12,13 @@
             return ConverterUtil.convertDigit(unitVal % 10) + " ";
         }
 
-        public override bool IsMatch(int num)
+        public override bool IsMatch(long num)
         {
             unitVal = num % 100;
             return unitVal > 20;
         }
 
-        public override bool InsertAnd(int num)
+        public override bool InsertAnd(long num)
         {
             return false;
         }
